requirements.txt
flask==2.2.0
flask-login
jinja2==3.1.2
psycopg2==2.8.4
flask-migrate==2.5.2
python-decouple
flask-sqlalchemy==2.5.1
werkzeug==2.2.2
pytz
python-dateutil
pandas
flask-mail
python-dotenv
pdfkit
wkhtmltopdf